requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# LLM Backends
openai>=1.3.0

# Caching
redis>=5.0.0

# Sentence Transformers (semantic cache)
sentence-transformers>=2.2.2

# Monitoring
prometheus-client>=0.19.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Security / Sanitization
bleach>=6.1.0

# Numeric / ML utilities
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Linting / Formatting
flake8>=6.1.0
black>=23.11.0

# Security scanning
bandit>=1.7.5

# Jupyter (demo notebook)
jupyter>=1.0.0
ipykernel>=6.26.0
matplotlib>=3.8.0
